pyproject.toml
[project]
name = "travel-assistant-rag-datadog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddtrace>=4.0.0",
    "fastapi>=0.123.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-qdrant>=0.2.1",
    "langfuse>=3.10.3",
    "langgraph>=1.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "sentence-transformers>=5.1.2",
    "uvicorn[standard]>=0.38.0",
]
